1st Version compatible with LS
* 1st working version that's fully compatible with the LiveScan
subclass. Only tested with running reconstruction using
"python *reconscript.py*" and "mpirun -n 1 python *reconscript.py*".

* Changed method of sending the load-reply from send_json() to
send_pyobj() and send() for the motor-message and the detector
image, respectively.

* Fixed indices used to retrieve the correct motor-messages and
detector images for the load-reply. The indices now works both
if the number available messages or if the number of images are
greater than the other.

* Makes sure that all frames have been recieved by ptypy before
closing the relay_socket, once ptypy has recieved the check-reply
with (end_of_scan and end_of_det_stream) = True, it requests a
'stop', and the RelayServer then sends back a dummy reply before
closing the relay_socket.