Update: preprocess
Changes to make the preprocessing as soon a a new frame comes in,
instead of waiting until the client requests to load the data.
So far only the cropping with a known center has been tested
and works independently of whether the preprocess information
from the client has been received before or after the first frames
have arrived.

Updates are still needed for:
* auto detection of center
* rebinning
* using a mask